+++ R/side_monad-to-monad.R
@@ -9,18 +9,6 @@ forget <- function(m){
   m
 }
 
-#' Add documentation to a monad
-#'
-#' @family monad-to-monad
-#' @param m An Rmonad
-#' @param ... any number of strings
-#' @return The input Rmonad with an added doc string
-#' @export
-doc <- function(m, ...){
-  m_doc(m) <- paste(list(...), collapse=" ")
-  m
-}
-
 #' Remove a level of nesting in an Rmonad
 #'
 #' @family monad-to-monad