+++ NEWS.md
@@ -9,6 +9,8 @@
  * Allow nesting of Rmonads, `as_monad` no longer automatically unnests them,
    there is an explicit function for this, `unnest`.
 
+ * Remove the `doc` function. Instead use doc strings.
+
 ## New features
 
  * Generation of rudimentary Markdown reports from pipelines with `mreport`